requirements.txt
pypdf2
python-docx
sentence-transformers
faiss-cpu
transformers
datasets
torch
pickle
fastapi
uvicorn
streamlit
requests